Add the un-namespaced wvui to LibUp's list of repos
Given this is an npm-based library that's going to be in the critical
path of painting (at least part of) almost every page in MediaWiki on
Wikimedia production sites, we really should be alert to security and
other issues as quickly as possible. Not quite sure whether this repo
will initially work nicely with LibUp; given its current likely 'non-
standard' configuration, probably not, but let's give it a whirl, and
thence iterate.

Bug: T257579
Change-Id: I9aa4726bacb2dd9ec8b8193f6605d74dfdb4287d