releases: Pin PHPUnit to 9.5.28 (latest 9.5.x as of today)
We probably should have been doing this already, but given the upstream
breaking change in 9.6.0 this has become acute.

There are a number of repos still on 8.5.x in their development branch,
which would fail to auto-upgrade (the process is sadly manual, see
T320245), so skip repos on that for now.

MW-style versioned repos need to stick to allowing both 8.5 and 9.5 in
their pre-1.40 branches (as that's when we landed eef1d7265), so have
them try for both but pinned to 9.5.28.

Bug: T328741
Change-Id: Id43019f0454d34fb28acc8d5a61d559f45b7a910